# Use BitVector128 in CharacterMap<T>