@KrisVandermotten 2020-02-05
I changed the implementation to use four uint values instead of two ulong values.

The performance is identical on X64, but I it can only be better on X86 or other 32 bit platforms.