@xoofx 2020-02-06
Nice trick, thanks!